+++ next.config.js
@@ -5,9 +5,18 @@ module.exports = {
     config.module.rules.push({
       test: /\.svg$/i,
       issuer: /\.[jt]sx?$/,
-      use: ['@svgr/webpack'],
-    })
+      use: ["@svgr/webpack"],
+    });
 
-    return config
+    return config;
   },
-}
+  async redirects() {
+    return [
+      {
+        source: "/",
+        destination: "/Dictionary",
+        permanent: true,
+      },
+    ];
+  },
+};
